+++ src/main/java/com/study/petory/domain/user/controller/TestSchedulerController.java
@@ -119,7 +119,7 @@ public class TestSchedulerController {
 		// 현재 날짜를 임의로 설정해서 테스트 (예: 휴면 계정으로 전환 예정 날짜인 "2025-06-18T00:00")
 		LocalDateTime simulatedNow = LocalDateTime.parse(date);
 
-		userDeactivationScheduler.testDeactivateInactiveUsers(simulatedNow);
+		userSchedulerService.testDeactivateInactiveUsers(simulatedNow);
 
 		return CommonResponse.of(SuccessCode.USER_DEACTIVATED);
 	}
